Update README with project description and technology stack. Move useValidation hook to hooks and update import path.
