+++ README.md
@@ -1,8 +1,27 @@
-# Final Project
+# Legacy Lock 
 
-Replace this readme with your own information about your project.
+Legacy Lock is a Virtual Time Capsule app that allows users to create and store messages or media (photos, videos, text) that are to be unlocked at a future date. 
 
-Start by briefly describing the assignment in a sentence or two. Keep it short and to the point.
+Authentication ensures that only authorized people can access the content, providing a secure and personal experience for users to preserve memories or send time-locked messages.
+
+## Technology Stack 
+
+Frontend: React
+Backend: Node.js with Express
+Database: MongoDB.
+Libraries: 
+  - React Router for navigation
+  - Zustand for global state management.
+React Hooks: useState, useEffect, useContext, and custom hooks to manage and validate user data and content.
+  ### 🧠 useMemo (in LatestLocket.jsx)
+  `useMemo` is a built-in React hook. 
+    * - useMemo stores the filtered and sorted capsule list to avoid recalculations on every render.
+    * - Filters out only future capsules (openAt > now)
+    * - Sorts capsules in ascending order (soonest first)
+
+External Libraries:
+date-fns for date handling and formatting
+multer for handling file uploads (for media files like photos and videos)
 
 ### 🧠 useMemo (in LatestLocket.jsx)
 `useMemo` is a built-in React hook. 
@@ -35,10 +54,6 @@ Start by briefly describing the assignment in a sentence or two. Keep it short a
 
   Read more: https://github.com/streamich/react-use/blob/HEAD/docs/useWindowSize.md
 
-## The problem
-
-Describe how you approached to problem, and what tools and techniques you used to solve it. How did you plan? What technologies did you use? If you had more time, what would be next?
-
 ## View it live
 
-Every project should be deployed somewhere. Be sure to include the link to the deployed project so that the viewer can click around and see what it's all about.
+[Legacy Locket App](https://legacy-locket.netlify.app/)
